mkmurn updated
mkmurn modified to alert the user if no energies are found in OUTCAR.
This can happen if one of the subdirectories contains a job that has
just restarted, overwriting an old OUTCAR, but hasn't yet finished an
ionic step. (previously this scenario caused a crash with a bunch of
ruby errors).